Split up GraphQL definition.
