+++ api/package.json
@@ -8,10 +8,12 @@
   },
   "dependencies": {
     "@babel/polyfill": "^7.0.0",
+    "apollo-server": "^2.2.2",
     "apollo-server-lambda": "^2.1.0",
     "dotenv": "^6.1.0",
     "graphql": "^0.13.0",
     "knex": "^0.15.2",
+    "lodash": "^4.17.11",
     "objection": "^1.3.0",
     "pg": "^7.5.0"
   },
