+++ api/src/functions/graphql.js
@@ -1,86 +1,39 @@
-import { ApolloServer, gql } from 'apollo-server-lambda';
+import { merge } from 'lodash';
+import { ApolloServer } from 'apollo-server-lambda';
+import { makeExecutableSchema } from 'apollo-server';
 
 import database from 'src/lib/database';
 
-import Category from 'src/models/Category';
-import Post from 'src/models/Post';
+import {
+  typeDefs as postTypeDefs,
+  resolvers as postResolvers,
+} from 'src/graphql/post.js';
+
+import {
+  typeDefs as categoryTypeDefs,
+  resolvers as categoryResolvers,
+} from 'src/graphql/category.js';
 
 database.init();
 
-// Construct a schema, using GraphQL schema language
-export const typeDefs = gql`
-  type Category {
-    id: ID!
-    name: String!
-    slug: String!
-    posts: [Post]
-  }
-
-  input CategoryInput {
-    name: String!
-    slug: String!
-  }
-
-  type Post {
-    id: ID!
-    slug: String!
-    title: String!
-    text: String!
-    category: Category
-  }
-
-  input PostInput {
-    title: String!
-    slug: String!
-    text: String!
-    categoryId: ID!
-  }
-
+const baseTypeDefs = `
   type Query {
-    post(slug: String!): Post
-    categories: [Category]
-    category(slug: String!): Category
+    _empty: String
   }
-
   type Mutation {
-    postCreate(post: PostInput!): Post
-    categoryCreate(category: CategoryInput!): Category
-    postAddCategory(postId: ID!, categoryId: ID!): Post
+    _empty: String
   }
 `;
 
-// TODO: Checkout https://github.com/vincit/objection-graphql#onquery
-export const resolvers = {
-  Query: {
-    post: (_, { slug }) => Post.query()
-      .eager('category')
-      .findOne({ slug }),
-    categories: () => Category.query().eager('posts'),
-    category: (_, { slug }) => Category.query()
-      .eager('posts')
-      .findOne({ slug }),
-  },
+const baseResolvers = {};
 
-  Mutation: {
-    postCreate: (_, args) => {
-      return Post.query().insert(args.post);
-    },
-    categoryCreate: (_, args) => {
-      return Category.query().insert(args.category);
-    },
-    postAddCategory: (_, args) => {
-      return Post.query().patchAndFetchById(args.postId, { categoryId: args.categoryId });
-    },
-  },
-
-  Category: {
-    posts: category => category.posts,
-  },
-};
+const schema = makeExecutableSchema({
+  typeDefs: [baseTypeDefs, postTypeDefs, categoryTypeDefs],
+  resolvers: merge(baseResolvers, postResolvers, categoryResolvers),
+});
 
 const server = new ApolloServer({
-  typeDefs,
-  resolvers,
+  schema,
 });
 
 export const handler = server.createHandler();
